+++ src/app/modules/master-data/components/types-of-companies/components/type-of-companies-modal/company-types-modal.component.scss
